requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
reportlab==4.0.4
Werkzeug==2.3.7
WTForms==3.0.1
gunicorn==21.2.0
requests==2.31.0
gpxpy==1.5.0
networkx==3.1
geopy==2.4.0
folium==0.14.0
PyJWT==2.8.0
email-validator==2.1.0
python-dotenv==1.0.0